+++ src/algorithm/math/missingNumber/missingNumber.ts
@@ -1,9 +1,13 @@
 export function missingNumber(nums: number[]): number {
   const n = nums.length;
-  let result = n;
+  let result = 0;
 
   for (let i = 0; i < n; i++) {
-    result ^= i ^ nums[i];
+    result ^= nums[i];
+  }
+
+  for (let i = 0; i <= n; i++) {
+    result ^= i;
   }
 
   return result;